+++ docs/config.rst
@@ -1,0 +1,4 @@
+.. show-options::
+
+	inmanta.server.config
+	inmanta.agent.config
