+++ src/inmanta/config.py
@@ -168,6 +168,23 @@ def is_uuid_opt(value):
 
 
 class Option(object):
+    """
+    Defines an option and exposes it for use
+
+    All config option should be define prior to use
+    For the document generator to work properly, they should be defined at the module level.
+
+    :param section: section in the config file
+    :param name: name of the option
+    :param default: default value for this option
+        the default value is either a value or a function.
+        If it is a value, `str(default)` will be used a default value.
+        If it is a function, its doc string will be used to represent the value in documentation.
+        and its return value as the actual default value
+    :param documentation: the documentation for this option
+    :param validator: a function responsible for turning the string representation of the option into the correct type. 
+        Its docstring is used as representation for the type of the option.
+    """
 
     def __init__(self, section, name, default, documentation, validator=is_str):
         self.section = section
@@ -206,6 +223,8 @@ class Option(object):
 
 #############################
 # Config
+#
+# Global config options are defined here
 #############################
 # flake8: noqa: H904
 state_dir = \
@@ -216,10 +235,6 @@ log_dir = \
     Option("config", "log_dir", "/var/log/inmanta",
            "The directory where the server stores log file. Currently this is only for the output of embedded agents.")
 
-heartbeat = \
-    Option("config", "heartbeat-interval", 10,
-           "Interval between subsequent heartbeats towards the server, lower number reduces load in the server", is_int)
-
 
 def get_default_nodename():
     """ socket.gethostname() """
@@ -228,25 +243,3 @@ def get_default_nodename():
 
 nodename = \
     Option("config", "node-name", get_default_nodename, "Force the hostname of this machine to a specific value", is_str)
-
-###############################
-# Transport Config
-###############################
-
-
-class TransportConfig(object):
-
-    def __init__(self, name):
-        self.prefix = "%s_rest_transport" % name
-        self.host = Option(self.prefix, "host", "localhost", "IP address or hostname of the server", is_str)
-        self.port = Option(self.prefix, "port", 8888, "Server port", is_int)
-        self.ssl = Option(self.prefix, "ssl", False, "Connect using SSL?", is_bool)
-        self.ssl_ca_cert_file = Option(
-            self.prefix, "ssl_ca_cert_file", None, "CA cert file used to validate the server certificate against", is_str_opt)
-        self.password = Option(
-            self.prefix, "password", None, "Password used to connect to the server", is_str_opt)
-        self.username = Option(
-            self.prefix, "username", None, "Username used to connect to the server", is_str_opt)
-
-TransportConfig("compiler")
-TransportConfig("client")
