+++ tests/test_2way_protocol.py
@@ -22,9 +22,10 @@ import uuid
 
 import colorlog
 from inmanta import methods
-from inmanta.config import Config, TransportConfig
+from inmanta.config import Config
 from tornado import gen
 from tornado.ioloop import IOLoop
+from inmanta.protocol import TransportConfig
 
 LOGGER = logging.getLogger(__name__)
